Import common.ver and ntverp.h from ReactOS to get dhahelper.rc to compile with MinGW. Compile dhahelper.rc and link it into dhahelper.sys on MinGW. Fix (harmless) "cast from pointer to integer of different size" warning in dhahelper.c.
git-svn-id: http://svn.tilp.info/repos/tilp@3704

+++ libticables/trunk/src/win32/dha/dhahelper.c
@@ -258,7 +258,7 @@ static STDCALL NTSTATUS MapPhysicalMemoryToLinearSpace(PVOID pPhysAddress,ULONG 
   PHYSICAL_ADDRESS   pStartPhysAddress;
   OutputDebugString ("dhahelper: entering MapPhysicalMemoryToLinearSpace");
     
-  pStartPhysAddress.QuadPart = (ULONGLONG)pPhysAddress;
+  pStartPhysAddress.QuadPart = (ULONGLONG)(ULONG)pPhysAddress;
 #ifndef NO_SEH
   __try {
 #endif
